+++ CHANGELOG.md
@@ -5,6 +5,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## [0.29.13] - 2025-09-15
+
 ### Added
 
 - Add `Diagnostics.energy_transfer_matrix`.